consolidate helper function files
